feat: 心愿单相关接口添加参数校验

+++ src/app/api/food.ts
@@ -1,7 +1,7 @@
 /*
  * @Author: Rock Chang
  * @Date: 2022-01-06 12:24:12
- * @LastEditTime: 2022-01-09 22:12:46
+ * @LastEditTime: 2022-01-10 14:34:47
  * @Description: 吃什么 - 菜品接口
  */
 import Router from 'koa-router';
@@ -11,8 +11,12 @@ import {
   SuccessResponse,
   ErrorResponse,
 } from '@/core/http-exception';
-import { PositiveIntValidator, Validator } from '@/app/validators/demo';
-import { ADMIN } from '@/constant/emun';
+import {
+  PositiveIntValidator,
+  NameValidator,
+  Validator,
+} from '@/app/validators/demo';
+import { NameIdValidator } from '@/app/validators/wish';
 import { API } from '@/constant/config';
 import { FoodModel } from '@/app/model/food';
 import { ERR_CODE } from '@/constant/emun';
@@ -24,7 +28,7 @@ router.prefix(`${API.PROJECT_INTERFACE_PREFIX}/chang/food`);
 
 // 新增菜品
 router.post('/add', new Auth().init, async (ctx: any) => {
-  const vs = await new Validator().validate(ctx);
+  const vs = await new NameValidator().validate(ctx);
   const {
     id,
     name,
@@ -49,16 +53,15 @@ router.post('/add', new Auth().init, async (ctx: any) => {
     material,
     process,
   });
-  // 存在id说明是jd的数据, 默认创建者1, 没有 id 则带上 uid
+  // 存在id说明是jd的数据, 默认创建者1
   if (id) {
-    // 判断要新增的是否存在
-    const AddRes = await FoodModel.getOne({
-      id,
-    });
+    // 判断要新增的是否存在, 存在直接返回
+    const AddRes = await FoodModel.getOne({ id });
     if (AddRes) throw new DataResponse(AddRes);
     // jd的数据 tag为string, 需转换为数组
     data.tag = tag ? (typeof tag === 'string' ? tag.split(',') : tag) : [];
     // if (AddRes) throw new ErrorResponse(ERR_CODE[5100]);
+    // 没有id说明是自建数据, 则带上 uid
   } else {
     data.uid = ctx.user.id;
   }
@@ -72,41 +75,27 @@ router.delete('/:id', new Auth().init, async (ctx: any) => {
   const { id } = vs.get('path');
   const user = ctx.user;
   // 判断要删除的是否存在
-  const res = await FoodModel.getOne({
-    uid: user.id,
-    id,
-  });
-  if (!res) throw new ErrorResponse(ERR_CODE[7]);
+  const res = await FoodModel.getOne({ uid: user.id, id });
+  if (!res) throw new ErrorResponse(ERR_CODE[5102]);
   // 删除
-  await FoodModel.destroy({
-    where: {
-      uid: user.id,
-      id,
-    },
-  });
+  await FoodModel.destroy({ where: { uid: user.id, id } });
   throw new SuccessResponse();
 });
 
 // 更新菜品
 router.put('/:id', new Auth().init, async (ctx: any) => {
-  const vs = await new PositiveIntValidator().validate(ctx);
+  const vs = await new NameIdValidator().validate(ctx);
   const { id } = vs.get('path');
   const { name, summary, tag, food_list } = vs.get('body');
   const data = removeEmpty({ name, summary, tag, food_list }); // 去除值为null|undefined的属性
-  const user = ctx.user;
-  // 判断要删除的是否存在
-  const deleteRes = await FoodModel.getOne({
-    uid: user.id,
-    id,
-  });
-  if (!deleteRes) throw new ErrorResponse(ERR_CODE[7]);
+  // 判断要更新的是否存在
+  const getRes = await FoodModel.getOne({ uid: ctx.user.id, id });
+  if (!getRes) throw new ErrorResponse(ERR_CODE[7]);
   // 更新
-  const updateRes = await FoodModel.update(data, {
-    where: { id },
-  });
+  const updateRes = await FoodModel.update(data, { where: { id } });
   if (updateRes[0] > 0) {
     // 更新成功, 返回新数据
-    const res = await FoodModel.getOne({ id });
+    const res = await FoodModel.findByPk(id);
     throw new DataResponse(res);
   } else {
     new ErrorResponse(ERR_CODE[8]);
